Move types, clean imports, move interpolate method
