+++ src/expression/expression.test.ts
@@ -3,8 +3,8 @@ import definitions from './definitions';
 import v8 from '../reference/v8.json' with {type: 'json'};
 import {createExpression, ICanonicalTileID, StyleExpression, StylePropertySpecification} from '..';
 import ParsingError from './parsing_error';
-import {VariableAnchorOffsetCollection} from './values';
 import {getGeometry} from '../../test/lib/geometry';
+import VariableAnchorOffsetCollection from './types/variable_anchor_offset_collection';
 
 // filter out internal "error" and "filter-*" expressions from definition list
 const filterExpressionRegex = /filter-/;
