+++ src/expression/types/padding.test.ts
@@ -24,4 +24,13 @@ describe('Padding', () => {
         const padding = new Padding([1, 2, 3, 4]);
         expect(padding.toString()).toBe('[1,2,3,4]');
     });
+
+    test('interpolate padding', () => {
+        const padding = new Padding([0, 0, 0, 0]);
+        const targetPadding = new Padding([1, 2, 6, 4]);
+
+        const i11nFn = (t: number) => Padding.interpolate(padding, targetPadding, t);
+        expect(i11nFn(0.5)).toBeInstanceOf(Padding);
+        expect(i11nFn(0.5)).toEqual(new Padding([0.5, 1, 3, 2]));
+    });
 });
