+++ src/expression/types/variable_anchor_offset_collection.test.ts
@@ -26,4 +26,18 @@ describe('VariableAnchorOffsetCollection', () => {
         const coll = new VariableAnchorOffsetCollection(['top', [2, 2]]);
         expect(coll.toString()).toBe('["top",[2,2]]');
     });
+
+    describe('interpolate variableAnchorOffsetCollection', () => {
+        const i11nFn = VariableAnchorOffsetCollection.interpolate;
+        const parseFn = VariableAnchorOffsetCollection.parse;
+
+        test('should throw with mismatched endpoints', () => {
+            expect(() => i11nFn(parseFn(['top', [0, 0]]), parseFn(['bottom', [1, 1]]), 0.5)).toThrow('Cannot interpolate values containing mismatched anchors. from[0]: top, to[0]: bottom');
+            expect(() => i11nFn(parseFn(['top', [0, 0]]), parseFn(['top', [1, 1], 'bottom', [2, 2]]), 0.5)).toThrow('Cannot interpolate values of different length. from: ["top",[0,0]], to: ["top",[1,1],"bottom",[2,2]]');
+        });
+
+        test('should interpolate offsets', () => {
+            expect(i11nFn(parseFn(['top', [0, 0], 'bottom', [2, 2]]), parseFn(['top', [1, 1], 'bottom', [4, 4]]), 0.5).values).toEqual(['top', [0.5, 0.5], 'bottom', [3, 3]]);
+        });
+    });
 });
