+++ src/expression/types/variable_anchor_offset_collection.ts
@@ -1,4 +1,6 @@
-import {VariableAnchorOffsetCollectionSpecification} from '../types.g';
+import RuntimeError from '../runtime_error';
+import {interpolateNumber} from '../../util/interpolate-primitives';
+import type {VariableAnchorOffsetCollectionSpecification} from '../../types.g';
 
 /** Set of valid anchor positions, as a set for validation */
 const anchors = new Set(['center', 'left', 'right', 'top', 'bottom', 'top-left', 'top-right', 'bottom-left', 'bottom-right']);
@@ -46,6 +48,32 @@ class VariableAnchorOffsetCollection {
 
     toString(): string {
         return JSON.stringify(this.values);
+    }
+
+    static interpolate(from: VariableAnchorOffsetCollection, to: VariableAnchorOffsetCollection, t: number): VariableAnchorOffsetCollection {
+        const fromValues = from.values;
+        const toValues = to.values;
+    
+        if (fromValues.length !== toValues.length) {
+            throw new RuntimeError(`Cannot interpolate values of different length. from: ${from.toString()}, to: ${to.toString()}`);
+        }
+    
+        const output: VariableAnchorOffsetCollectionSpecification = [];
+    
+        for (let i = 0; i < fromValues.length; i += 2) {
+            // Anchor entries must match
+            if (fromValues[i] !== toValues[i]) {
+                throw new RuntimeError(`Cannot interpolate values containing mismatched anchors. from[${i}]: ${fromValues[i]}, to[${i}]: ${toValues[i]}`);
+            }
+            output.push(fromValues[i]);
+    
+            // Interpolate the offset values for each anchor
+            const [fx, fy] = fromValues[i + 1] as [number, number];
+            const [tx, ty] = toValues[i + 1] as [number, number];
+            output.push([interpolateNumber(fx, tx, t), interpolateNumber(fy, ty, t)]);
+        }
+    
+        return new VariableAnchorOffsetCollection(output);
     }
 }
