+++ src/function/index.ts
@@ -1,13 +1,12 @@
-import Color from '../util/color';
+import Color, {isSupportedInterpolationColorSpace} from '../expression/types/color';
 import extend from '../util/extend';
 import getType from '../util/get_type';
-import interpolate, {isSupportedInterpolationColorSpace} from '../util/interpolate';
-import Interpolate from '../expression/definitions/interpolate';
+import Interpolate, { interpolateFactory } from '../expression/definitions/interpolate';
 import Formatted from '../expression/types/formatted';
 import ResolvedImage from '../expression/types/resolved_image';
 import {supportsInterpolation} from '../util/properties';
 import {findStopLessThanOrEqualTo} from '../expression/stops';
-import Padding from '../util/padding';
+import Padding from '../expression/types/padding';
 
 export function isFunction(value) {
     return typeof value === 'object' && value !== null && !Array.isArray(value);
@@ -174,7 +173,7 @@ function evaluateExponentialFunction(parameters, propertySpec, input) {
 
     const outputLower = parameters.stops[index][1];
     const outputUpper = parameters.stops[index + 1][1];
-    const interp = interpolate[propertySpec.type] || identityFunction;
+    const interp = interpolateFactory[propertySpec.type] || identityFunction;
 
     if (typeof outputLower.evaluate === 'function') {
         return {
